# Optimize #at_css and #css initialization