--- a/lib/nokolexbor/node.rb
+++ b/lib/nokolexbor/node.rb
@@ -717,18 +717,38 @@ module Nokolexbor
 
     def xpath_query_from_css_rule(rule, ns)
       ensure_nokogiri
-      if defined? Nokogiri::CSS::XPathVisitor::BuiltinsConfig
-        visitor = Nokogiri::CSS::XPathVisitor.new(
-          builtins: Nokogiri::CSS::XPathVisitor::BuiltinsConfig::OPTIMAL,
-          doctype: :html4,
-        )
-      else
-        visitor = Nokogiri::CSS::XPathVisitorOptimallyUseBuiltins.new
+
+      unless defined?(Gem)
+        require 'rubygems'
       end
-      self.class::IMPLIED_XPATH_CONTEXTS.map do |implied_xpath_context|
-        Nokogiri::CSS.xpath_for(rule.to_s, { prefix: implied_xpath_context, ns: ns,
-                                   visitor: visitor, })
-      end.join(" | ")
+
+      v_1_17_0 = Gem::Version.new("1.17.0")
+      current_version = Gem::Version.new(Nokogiri::VERSION)
+
+      if current_version < v_1_17_0
+        if defined? Nokogiri::CSS::XPathVisitor::BuiltinsConfig
+          visitor = Nokogiri::CSS::XPathVisitor.new(
+            builtins: Nokogiri::CSS::XPathVisitor::BuiltinsConfig::OPTIMAL,
+            doctype: :html4,
+          )
+        else
+          visitor = Nokogiri::CSS::XPathVisitorOptimallyUseBuiltins.new
+        end
+        self.class::IMPLIED_XPATH_CONTEXTS.map do |implied_xpath_context|
+          Nokogiri::CSS.xpath_for(rule.to_s, { prefix: implied_xpath_context, ns: ns,
+                                    visitor: visitor, })
+        end.join(" | ")
+      else
+        self.class::IMPLIED_XPATH_CONTEXTS.map do |implied_xpath_context|
+          visitor = Nokogiri::CSS::XPathVisitor.new(
+            builtins: Nokogiri::CSS::XPathVisitor::BuiltinsConfig::OPTIMAL,
+            doctype: :html4,
+            prefix: implied_xpath_context,
+            namespaces: ns,
+          )
+          Nokogiri::CSS.xpath_for(rule.to_s, visitor: visitor)
+        end.join(" | ")
+      end
     end
 
     def extract_params(params)