--- a/lib/nokolexbor/version.rb
+++ b/lib/nokolexbor/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Nokolexbor
-  VERSION = '0.5.4'
+  VERSION = '0.5.5'
 end