--- a/patches/0004-lexbor-fix-template-clone.patch
+++ b/patches/0004-lexbor-fix-template-clone.patch
@@ -16,7 +16,7 @@ index a2153f4..8a9c69f 100755
  
 +        if (curr->local_name == LXB_TAG_TEMPLATE && curr->first_child != NULL && cnode->type == LXB_DOM_NODE_TYPE_DOCUMENT_FRAGMENT) {
 +            lxb_dom_node_remove(curr->first_child);
-+            lxb_html_interface_template(curr)->content = cnode;
++            lxb_html_interface_template(curr)->content = (lxb_dom_document_fragment_t *)cnode;
 +        }
 +
          lxb_dom_node_insert_child(curr, cnode);