--- a/spec/multi_thread_spec.rb
+++ b/spec/multi_thread_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe "Multi threading" do
   it 'should be thread-safe' do
-    10000.times.map do
+    1000.times.map do
       Thread.new do
         random_str = (0...32).map { (65 + rand(58)).chr }.join
         doc = Nokolexbor::HTML('<div><a class="a">' + random_str + '</a></div>')