Before
```
Nokolexbor at_css:   138765.3 i/s
Nokolexbor    css:    10132.0 i/s
```

After
```
Nokolexbor at_css:   202963.7 i/s
Nokolexbor    css:     10209.4 i/s
```