@zyc9012 2024-09-04
Thanks @Freaky. I've removed all pthread stuff and made sure that it's thread-safe on truffleruby.